+++ src/main/java/ua/cc/spon/shlibot/ShoppingListTelegramBotApplication.java
@@ -5,9 +5,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class ShoppingListTelegramBotApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ShoppingListTelegramBotApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ShoppingListTelegramBotApplication.class, args);
+    }
 
 }