+++ src/main/java/ua/cc/spon/shlibot/service/ProductServiceImpl.java
@@ -4,6 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ua.cc.spon.shlibot.repository.ProductRepository;
 import ua.cc.spon.shlibot.repository.entity.Product;
+import ua.cc.spon.shlibot.repository.entity.UserList;
+
+import java.util.List;
+import java.util.Optional;
 
 /**
  * Implementation of {@link ProductService}.
@@ -24,4 +28,18 @@ public class ProductServiceImpl implements ProductService {
         productRepository.save(product);
     }
 
+    @Override
+    public List<Product> getAllByListIs(UserList list) {
+        return productRepository.getAllByListIs(list);
+    }
+
+    @Override
+    public Optional<Product> findById(int id) {
+        return productRepository.findById(id);
+    }
+
+    @Override
+    public Product deleteById(int id) {
+        return productRepository.deleteById(id);
+    }
 }