+++ src/main/java/ua/cc/spon/shlibot/service/UserListService.java
@@ -3,6 +3,7 @@ package ua.cc.spon.shlibot.service;
 import ua.cc.spon.shlibot.repository.entity.TelegramUser;
 import ua.cc.spon.shlibot.repository.entity.UserList;
 
+import java.util.Optional;
 import java.util.Set;
 
 /**
@@ -13,10 +14,12 @@ public interface UserListService {
     /**
      * Save provided {@link UserList} entity.
      *
-     * @param  userList provided telegram user.
+     * @param userList provided telegram user.
      */
     void save(UserList userList);
 
     Set<UserList> findByOwner(TelegramUser telegramUser);
+
+    Optional<UserList> findById(int id);
 
 }