+++ src/main/resources/db/migration/V00007__add_tg_user_table.sql
@@ -1,0 +1,4 @@
+alter table tg_user
+    add constraint tg_user_lists_fkey
+        foreign key (mainlist_id) references lists (list_id)
+            on delete set null;